fix: Allow groups to have admin permission
